refactor: move isReplicaSetReferenced to replicaset.go
