+++ rollout/canary.go
@@ -4,7 +4,6 @@ import (
 	"sort"
 
 	appsv1 "k8s.io/api/apps/v1"
-	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/utils/pointer"
 
@@ -15,7 +14,6 @@ import (
 	"github.com/argoproj/argo-rollouts/utils/record"
 	replicasetutil "github.com/argoproj/argo-rollouts/utils/replicaset"
 	rolloututil "github.com/argoproj/argo-rollouts/utils/rollout"
-	serviceutil "github.com/argoproj/argo-rollouts/utils/service"
 )
 
 func (c *rolloutContext) rolloutCanary() error {
@@ -240,59 +238,6 @@ func (c *rolloutContext) scaleDownOldReplicaSetsForCanary(oldRSs []*appsv1.Repli
 	}
 
 	return totalScaledDown, nil
-}
-
-// isReplicaSetReferenced returns if the given ReplicaSet is still being referenced by any of
-// the current, stable, blue-green services. Used to determine if the ReplicaSet can
-// safely be scaled to zero, or deleted.
-func (c *rolloutContext) isReplicaSetReferenced(rs *appsv1.ReplicaSet) bool {
-	rsPodHash := replicasetutil.GetPodTemplateHash(rs)
-	if rsPodHash == "" {
-		return false
-	}
-	ro := c.rollout
-	referencesToCheck := []string{
-		ro.Status.StableRS,
-		ro.Status.CurrentPodHash,
-		ro.Status.BlueGreen.ActiveSelector,
-		ro.Status.BlueGreen.PreviewSelector,
-	}
-	if ro.Status.Canary.Weights != nil {
-		referencesToCheck = append(referencesToCheck, ro.Status.Canary.Weights.Canary.PodTemplateHash, ro.Status.Canary.Weights.Stable.PodTemplateHash)
-	}
-	for _, ref := range referencesToCheck {
-		if ref == rsPodHash {
-			return true
-		}
-	}
-
-	// The above are static, lightweight checks to see if the selectors we record in our status are
-	// still referencing the ReplicaSet in question. Those checks aren't always enough. Next, we do
-	// a deeper check to look up the actual service objects, and see if they are still referencing
-	// the ReplicaSet. If so, we cannot scale it down.
-	var servicesToCheck []string
-	if ro.Spec.Strategy.Canary != nil {
-		servicesToCheck = []string{ro.Spec.Strategy.Canary.CanaryService, ro.Spec.Strategy.Canary.StableService}
-	} else {
-		servicesToCheck = []string{ro.Spec.Strategy.BlueGreen.ActiveService, ro.Spec.Strategy.BlueGreen.PreviewService}
-	}
-	for _, svcName := range servicesToCheck {
-		if svcName == "" {
-			continue
-		}
-		svc, err := c.servicesLister.Services(c.rollout.Namespace).Get(svcName)
-		if err != nil {
-			if k8serrors.IsNotFound(err) {
-				// service doesn't exist
-				continue
-			}
-			return true
-		}
-		if serviceutil.GetRolloutSelectorLabel(svc) == rsPodHash {
-			return true
-		}
-	}
-	return false
 }
 
 // isDynamicallyRollingBackToStable returns true if we were in the middle of an canary update with
