+++ rollout/replicaset_test.go
@@ -8,6 +8,9 @@ import (
 	"github.com/stretchr/testify/assert"
 	appsv1 "k8s.io/api/apps/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/util/intstr"
+	k8sinformers "k8s.io/client-go/informers"
 	k8sfake "k8s.io/client-go/kubernetes/fake"
 	core "k8s.io/client-go/testing"
 	"k8s.io/utils/pointer"
@@ -350,6 +353,185 @@ func TestReconcileOldReplicaSet(t *testing.T) {
 				t.Errorf("expected scaling to occur")
 				return
 			}
+		})
+	}
+}
+
+func TestIsReplicaSetReferenced(t *testing.T) {
+	newRSWithPodTemplateHash := func(hash string) *appsv1.ReplicaSet {
+		return &appsv1.ReplicaSet{
+			ObjectMeta: metav1.ObjectMeta{
+				Labels: map[string]string{
+					v1alpha1.DefaultRolloutUniqueLabelKey: hash,
+				},
+			},
+		}
+	}
+
+	testCases := []struct {
+		name           string
+		status         v1alpha1.RolloutStatus
+		canaryService  string
+		stableService  string
+		activeService  string
+		previewService string
+		services       []runtime.Object
+		rsHash         string
+		expectedResult bool
+	}{
+		{
+			name:           "empty hash",
+			status:         v1alpha1.RolloutStatus{StableRS: "abc123"},
+			rsHash:         "",
+			expectedResult: false,
+		},
+		{
+			name:           "not referenced",
+			status:         v1alpha1.RolloutStatus{StableRS: "abc123"},
+			rsHash:         "def456",
+			expectedResult: false,
+		},
+		{
+			name:           "stable rs referenced",
+			status:         v1alpha1.RolloutStatus{StableRS: "abc123"},
+			rsHash:         "abc123",
+			expectedResult: true,
+		},
+		{
+			name:           "current rs referenced",
+			status:         v1alpha1.RolloutStatus{CurrentPodHash: "abc123"},
+			rsHash:         "abc123",
+			expectedResult: true,
+		},
+		{
+			name:           "active referenced",
+			status:         v1alpha1.RolloutStatus{BlueGreen: v1alpha1.BlueGreenStatus{ActiveSelector: "abc123"}},
+			rsHash:         "abc123",
+			expectedResult: true,
+		},
+		{
+			name:           "active referenced",
+			status:         v1alpha1.RolloutStatus{BlueGreen: v1alpha1.BlueGreenStatus{PreviewSelector: "abc123"}},
+			rsHash:         "abc123",
+			expectedResult: true,
+		},
+		{
+			name: "traffic routed canary current pod hash",
+			status: v1alpha1.RolloutStatus{Canary: v1alpha1.CanaryStatus{Weights: &v1alpha1.TrafficWeights{
+				Canary: v1alpha1.WeightDestination{
+					PodTemplateHash: "abc123",
+				},
+			}}},
+			rsHash:         "abc123",
+			expectedResult: true,
+		},
+		{
+			name: "traffic routed canary current pod hash",
+			status: v1alpha1.RolloutStatus{Canary: v1alpha1.CanaryStatus{Weights: &v1alpha1.TrafficWeights{
+				Stable: v1alpha1.WeightDestination{
+					PodTemplateHash: "abc123",
+				},
+			}}},
+			rsHash:         "abc123",
+			expectedResult: true,
+		},
+		{
+			name: "canary service still referenced",
+			status: v1alpha1.RolloutStatus{
+				CurrentPodHash: "abc123",
+				StableRS:       "abc123",
+			},
+			canaryService:  "mysvc",
+			services:       []runtime.Object{newService("mysvc", 80, map[string]string{v1alpha1.DefaultRolloutUniqueLabelKey: "def456"}, nil)},
+			rsHash:         "def456",
+			expectedResult: true,
+		},
+		{
+			name: "stable service still referenced",
+			status: v1alpha1.RolloutStatus{
+				CurrentPodHash: "abc123",
+				StableRS:       "abc123",
+			},
+			stableService:  "mysvc",
+			services:       []runtime.Object{newService("mysvc", 80, map[string]string{v1alpha1.DefaultRolloutUniqueLabelKey: "def456"}, nil)},
+			rsHash:         "def456",
+			expectedResult: true,
+		},
+		{
+			name: "active service still referenced",
+			status: v1alpha1.RolloutStatus{
+				CurrentPodHash: "abc123",
+				StableRS:       "abc123",
+			},
+			activeService:  "mysvc",
+			services:       []runtime.Object{newService("mysvc", 80, map[string]string{v1alpha1.DefaultRolloutUniqueLabelKey: "def456"}, nil)},
+			rsHash:         "def456",
+			expectedResult: true,
+		},
+		{
+			name: "preview service still referenced",
+			status: v1alpha1.RolloutStatus{
+				CurrentPodHash: "abc123",
+				StableRS:       "abc123",
+			},
+			activeService:  "mysvc",
+			previewService: "mysvc2",
+			services:       []runtime.Object{newService("mysvc2", 80, map[string]string{v1alpha1.DefaultRolloutUniqueLabelKey: "def456"}, nil)},
+			rsHash:         "def456",
+			expectedResult: true,
+		},
+		{
+			name: "service not found",
+			status: v1alpha1.RolloutStatus{
+				CurrentPodHash: "abc123",
+				StableRS:       "abc123",
+			},
+			activeService:  "mysvc",
+			previewService: "mysvc2",
+			rsHash:         "def456",
+			expectedResult: false,
+		},
+	}
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+
+			fake := fake.Clientset{}
+			k8sfake := k8sfake.NewSimpleClientset(tc.services...)
+			informers := k8sinformers.NewSharedInformerFactory(k8sfake, 0)
+			servicesLister := informers.Core().V1().Services().Lister()
+			stopchan := make(chan struct{})
+			defer close(stopchan)
+			informers.Start(stopchan)
+			informers.WaitForCacheSync(stopchan)
+
+			var r *v1alpha1.Rollout
+			if tc.activeService != "" {
+				r = newBlueGreenRollout("test", 1, nil, tc.activeService, tc.previewService)
+			} else {
+				r = newCanaryRollout("test", 1, nil, nil, nil, intstr.FromInt(0), intstr.FromInt(1))
+				r.Spec.Strategy.Canary.CanaryService = tc.canaryService
+				r.Spec.Strategy.Canary.StableService = tc.stableService
+			}
+			r.Status = tc.status
+
+			roCtx := &rolloutContext{
+				rollout: r,
+				log:     logutil.WithRollout(r),
+				reconcilerBase: reconcilerBase{
+					servicesLister:    servicesLister,
+					argoprojclientset: &fake,
+					kubeclientset:     k8sfake,
+					recorder:          record.NewFakeEventRecorder(),
+				},
+			}
+			rs := newRSWithPodTemplateHash(tc.rsHash)
+			stillReferenced := roCtx.isReplicaSetReferenced(rs)
+
+			assert.Equal(
+				t,
+				tc.expectedResult,
+				stillReferenced,
+			)
 		})
 	}
 }
